limit for fee collection
fees are lopped and adjusted with the vesting schedule.

only users with fees can be passed to input array.

we can also have an additional view function to list the users with fees>0 and then use that call first to generate the input array before actually interacting with the contract.